Add rudimentary support for finding functions.
If the user indicates that they want information about the functions
that wrap a given address (by constructing Mapping with
::with_functions), we also scan all DWARF subprograms and inlined
subroutines under each compilation unit. For each one, we store its
range and its name (linkage_name -> name -> abstract_origin).

In locate, if a unit holds any programs (which it only will when the
Mapping was created using with_functions), the programs for that unit
are searched to find the best fitting subprogram.

The programs are currently stored in a vector inside each unit, which is
fine if the number of subprograms for any given unit is relatively
small, but might pose a performance problem for larger applications. In
the future, we may want to move to a smarter data structure here.

This patch also adds a dependency on the git version of gimli so that we
can use entries_at_offset to resolve abstract origins. Once a new
release of gimli is cut, this can be removed.