Add first integration test.
We compare our output to that of addr2line (or atos on macOS), and check
that a) we locate at least some addresses successfully, and b) whenever
we do give a file/lineno, we give the same answer as addr2line/atos.

Note that atos on macOS is pretty stupid. First, it doesn't allow us to
pipe input to it and read the entries one by one, because it doesn't
flush its output for each input address. This means we have to collect
all the addresses, write them all to atos, close stdin, and then read
out the answers one by one. Second, it only outputs the file name, *not*
the full path, so we *could* have false positives for source files with
the same file name on macOS, and the tests wouldn't detect it.

macOS tests are currently disabled, as we produce line numbers that are
off-by-one in a number of places on rust nightly and beta (but not
stable). Fixing this is left for future work.