Build full-stack error handling with error-chain
This removes all unsafe unwrap/expect calls, and presents the user with
a consistent error interface that also gives the underlying causes where
possible.

There is one top-level unwrap() in Mapping::locate() since we would
probably rather return None to the user in that case. We could
eventually expose errors here, but providing a simplified API might be a
good idea?