Implement lazy binary search for line rows
Note that we skip line row sequences with addresses that are not higher
than those in previous sequences, as this would make the binary search
all wrong. This actually also impacts the current (non-binary) search,
in that it might terminate too early. Consider the case where we're
looking for addr 6. We might see:

 1 2 3 4 8 9 5 6

Both the old algorithm and this one would yield "4". Fixing this will be
the subject of a different PR.