Allow running benchmark.sh while repo is dirty
This stashes all changes before running the benchmark, and then pops
those changes again (if there were any) after all the binaries have been
built. Note that we also need to do a some trickery here to make sure
that the bash script doesn't do weird things because we modified it
while it was executing (which would happen if benchmark.sh itself was
dirty).

This patch is best viewed with git diff -w