Handle units without address ranges in .debug_info
Furthermore, ignore address ranges completely, because they are unreliable.
I have a shared library that was compiled with clang version 3.7.1
which has invalid address ranges such as:
    [14] range entry    0x00000001 0x00000001
The line table contains rows for the addresses that should have been here,
so we now ignore the address range in .debug_info and use only the line table
instead.

Fixes #42