Trim dependencies of `addr2line` for `backtrace`
This commit trims the dependencies of `addr2line` to be used in the
`backtrace` crate. The purpose here is to slowly trim down the list of
dependencies to make it easier to eventually include this in the
standard library. Fow now this handles three previously required
dependenies of `addr2line`:

* `smallvec` - this is made optional where if not present a standard
  `alloc::vec::Vec<T>` is used.
* `fallible-iterator` - this is simply made optional.
* `lazycell` - this is dropped in favor of a small hand-rolled
  implementation.

The new features are enabled by default and the thinking is that the
`backtrace` crate would disable all these features by default. If the
`std` feature of `backtrace` is activated it'd activate these features,
however.