Remove BTreeMap and go back to a Vec.
This saves the map lookup when querying functions.
I'm also removing the FrameIter's LocationOnly state again, and OuterFunction::parse
no longer returns an Option<>. If we've seen an entry for the function when building
the outer functions list, then we know the entry exists, and we don't need to handle
entries.is_empty() at that dw_die_offset.