Get unit ranges from the line program.
For units which don't declare any address ranges, fall back to
parsing the line program eagerly and then use the address ranges
of the line sequences in that line program to create unit ranges.

This helps with the .so files which are generated by perf inject --jit;
these files do not declare any unit ranges.