Support LineInstruction::DefineFile.
We were getting the file names from the wrong header.
As the line program rows are being iterated, we can add files
to the program header file list. However, we weren't recognizing
those added files because we were looking at the pre-clone ilnp
header, whereas the new files end up on the cloned header that is
stored in the rows.

On the .so file produced by perf inject --jit, this meant that we
ended up with a None filename.