Support relative file paths.
In the "jitted-103168-33.so" file that I was testing with, we were
getting the file path "/jitdump-cacheop-103168.txt" instead of just
the relative path "jitdump-cacheop-103168.txt".

The unit in this file has dw_unit.comp_dir set to None, and that unit
line program's only file has a directory_index() of zero.