Support split DWARF.
This changes most of the addr2line API entry points to take a continuation-
based approach that returns either the result or a continuation that can be
resumed once the appropriate split DWARF file is loaded. Actually loading
the split DWARF is deferred to the caller, just as it is for the main and
supplementary debug files.

Because the `gimli::Dwarf` for split DWARF varies by unit, the `Context::dwarf`
accessor is now gone. Instead `Context::find_dwarf_and_unit` returns the
correct `gimli::Dwarf` and `gimli::Unit` pair for a given address. The
`find_frames` API is more or less the same, just in continuation form now.

Callers that do not care to support split DWARF can simply import the
`LookupResultExt` trait and call `skip_all_loads` on the results of
continuation-based APIs to skip loading any split DWARF and immediately
retrieve whatever value can be produced in the absence of that split DWARF.

Internally there have been significant refactorings.

The existing treatment of a supplementary file as a full `ResDwarf` is
unnecessary. Supplementary files are only relevant to addr2line because they
may provide function names. Supplementary files by definition *cannot*
contain any information related to the layout of the binary, since they
are shared among multiple binaries. (In other words the `unit_ranges`
for a supplementary file is always empty). In practice the full `ResUnit`
for a partial compliation units in the supplementary file is also unnecessary,
though that's not changed here.

Split the parsed components from the gimli components, effectively undoing
the introduction of `ResDwarf`. The parsed components for the primary debug
file live in `ParsedDwarf`, along with an optional vector of `ResUnit`s from
the supplementary file. The gimli components now live in `RawDwarf`.

This allows `RawDwarf` to be trivially clonable, and `ParsedDwarf` is completely
independent of anything that would be found in a DWO/DWP.
`ResUnit::dwarf_and_unit_dwo` handles all of the details of invoking the loader
when appropriate and caching its results. Code paths that are going to examine
data that is affected by split DWARF (e.g. anything with functions) call this
function to get the correct `RawDwarf` and `gimli::Unit` to use.

The MSRV is now 1.55.0.