Enable builtin_split_dwarf_loader for Windows
I was unable to generate DWARF with mingw that had non-UTF8 paths.
Somewhere in the toolchain the path is being lossily converted to UTF8.
We may find a situation in which the DWARF doesn't contain UTF8 in the
future, but until then we try to do a lossless conversion to UTF8, and
return an error if that fails.