Store boxed slices instead of Vec objects in Context
We never resize the unit ranges, units, or sup units that we parsed --
they stay what they were when we instantiated the Context object. As
such, we don't need to store Vec objects, which include a capacity. Do
what we do for the Functions type and just store boxed slices. Doing so
cuts a machine word of each, but also trims heap allocations to the
minimal size required in the process.

Signed-off-by: Daniel Müller <deso@posteo.net>