Don't parse ranges for partial units
These don't normally have ranges anyway, but it is possible for
them to have a DW_AT_stmt_list that points to the same program
that contains ranges for a compilation unit, and loading the ranges
from that is wrong.

This used to work because our search would find the compilation unit
first, but we changed the search order and now it was finding
the partial unit.