Add tests for tombstones
This covers some but not all cases of tombstones.
In particular, we don't have a test for a DW_LNE_set_address
tombstone in the middle of a line number sequence.

This also handles an integer overflow for a tombstone address
in DW_AT_low_pc. We already did the right thing in release builds.