bin: prefer symbol table names over DWARF names
Prefer the symbol table over the DWARF name because:
- the symbol can include a clone suffix
- llvm may omit the linkage name in the DWARF with -g1

These two cases are tested in testoutput/dwarf/cold
and testoutput/dwarf/base-clang-g1.

This also fixes the case where the DWARF includes an inlined
function with no name, and we wrongly used the symbol table.
I don't have a test case for this.