Fix line parsing for split DWARF (#353)
The line number info is not contained in the DWO, so when parsing
it we need to use ctx.sections, not the DWO sections.

This was causing an error when we tried to finish parsing a line
program that used a .debug_line_str offset for a filename, and we
tried to read at that offset in the wrong .debug_line_str section.