Use unstable instead of stable sorts (#358)
* Use unstable instead of stable sorts

Unstable sorting has significantly better code size (and often better
performance too). With this change (rebased on top of v0.25.0, which is
what std currently uses) the .text size of Rust hello world (with `-Clto=fat
-Copt-level=s`) goes from 297120 to 290351 (-2.23%).

I do not know *that* much about addr2line, so I cannot guarantee that
this code is fine with unstable sorting. But I think it should be fine
since it always sorts on addresses, which are either unique or
nondeterministic anyways (as mentioned in one comment even). So it
should not make a significant difference.

* Sort parsed unit ranges by `unit_id` as well

Not necessary for correctness, but helps compare results.